+++ src/Context.php
@@ -9,15 +9,6 @@ class Context implements \ArrayAccess
     /**
      * @param array<(int|string), mixed> $data        The input parameter accepts an array or an Object
      * @param array<string, string>      $environment A list of environment variables to add to the command
-     * @param array{
-     *     'port'?: int,
-     *     'path_private_key'?: string,
-     *     'jump_host'?: string,
-     *     'multiplexing_control_path'?: string,
-     *     'multiplexing_control_persist'?: string,
-     *     'enable_strict_check'?: bool,
-     *     'password_authentication'?: bool,
-     * }                                 $sshOptions  Options for SSH connection
      */
     public function __construct(
         public readonly array $data = [],
@@ -30,9 +21,6 @@ class Context implements \ArrayAccess
         public readonly bool $allowFailure = false,
         public readonly bool $notify = false,
         public readonly VerbosityLevel $verbosityLevel = VerbosityLevel::NOT_CONFIGURED,
-        public readonly string $host = 'localhost',
-        public readonly ?string $user = null,
-        public readonly array $sshOptions = [],
     ) {
         $this->currentDirectory = $currentDirectory ?? PathHelper::getRoot();
     }
@@ -51,9 +39,6 @@ class Context implements \ArrayAccess
             $this->allowFailure,
             $this->notify,
             $this->verbosityLevel,
-            $this->host,
-            $this->user,
-            $this->sshOptions,
         );
     }
 
@@ -71,9 +56,6 @@ class Context implements \ArrayAccess
             $this->allowFailure,
             $this->notify,
             $this->verbosityLevel,
-            $this->host,
-            $this->user,
-            $this->sshOptions,
         );
     }
 
@@ -90,9 +72,6 @@ class Context implements \ArrayAccess
             $this->allowFailure,
             $this->notify,
             $this->verbosityLevel,
-            $this->host,
-            $this->user,
-            $this->sshOptions,
         );
     }
 
@@ -109,9 +88,6 @@ class Context implements \ArrayAccess
             $this->allowFailure,
             $this->notify,
             $this->verbosityLevel,
-            $this->host,
-            $this->user,
-            $this->sshOptions,
         );
     }
 
@@ -128,9 +104,6 @@ class Context implements \ArrayAccess
             $this->allowFailure,
             $this->notify,
             $this->verbosityLevel,
-            $this->host,
-            $this->user,
-            $this->sshOptions,
         );
     }
 
@@ -147,9 +120,6 @@ class Context implements \ArrayAccess
             $this->allowFailure,
             $this->notify,
             $this->verbosityLevel,
-            $this->host,
-            $this->user,
-            $this->sshOptions,
         );
     }
 
@@ -166,9 +136,6 @@ class Context implements \ArrayAccess
             $this->allowFailure,
             $this->notify,
             $this->verbosityLevel,
-            $this->host,
-            $this->user,
-            $this->sshOptions,
         );
     }
 
@@ -185,9 +152,6 @@ class Context implements \ArrayAccess
             $allowFailure,
             $this->notify,
             $this->verbosityLevel,
-            $this->host,
-            $this->user,
-            $this->sshOptions,
         );
     }
 
@@ -204,9 +168,6 @@ class Context implements \ArrayAccess
             $this->allowFailure,
             $notify,
             $this->verbosityLevel,
-            $this->host,
-            $this->user,
-            $this->sshOptions,
         );
     }
 
@@ -223,58 +184,6 @@ class Context implements \ArrayAccess
             $this->allowFailure,
             $this->notify,
             $verbosityLevel,
-            $this->host,
-            $this->user,
-            $this->sshOptions,
-        );
-    }
-
-    public function withoutSsh(): self
-    {
-        return new self(
-            $this->data,
-            $this->environment,
-            $this->currentDirectory,
-            $this->tty,
-            $this->pty,
-            $this->timeout,
-            $this->quiet,
-            $this->allowFailure,
-            $this->notify,
-            $this->verbosityLevel,
-            'localhost',
-            null,
-            [],
-        );
-    }
-
-    /**
-     * @param array{
-     *     'port'?: int,
-     *     'path_private_key'?: string,
-     *     'jump_host'?: string,
-     *     'multiplexing_control_path'?: string,
-     *     'multiplexing_control_persist'?: string,
-     *     'enable_strict_check'?: bool,
-     *     'password_authentication'?: bool,
-     * } $sshOptions Options for SSH connection
-     */
-    public function withSsh(string $host, string $user, array $sshOptions = []): self
-    {
-        return new self(
-            $this->data,
-            $this->environment,
-            $this->currentDirectory,
-            $this->tty,
-            $this->pty,
-            $this->timeout,
-            $this->quiet,
-            $this->allowFailure,
-            $this->notify,
-            $this->verbosityLevel,
-            $host,
-            $user,
-            $sshOptions,
         );
     }